refactor: Gauges Filter component
