+++ .vscode/settings.json
@@ -49,6 +49,7 @@
     "wagmi",
     "WBETH",
     "wbnb",
+    "ZKEVM",
     "ZKSYNC"
   ]
 }
